afl: loader: prevent waitpid() hang by using sigtimedwait.

+++ kernel_fuzzing/execve/loader.c
@@ -1,6 +1,8 @@
 #include <unistd.h>
 #include <fcntl.h>
 #include <sys/stat.h>
+#include <sys/types.h>
+#include <sys/wait.h>
 #include <signal.h>
 
 #include "config.h"
@@ -14,9 +16,36 @@ int load_hook(unsigned int argc, char** argv)
   null_fd = open("/dev/null", O_RDWR);
 }
 
+static sigset_t old_set;
+void pre_hook(unsigned argc, char** argv)
+{
+  sigset_t set;
+
+  child = -1;
+
+  sigemptyset(&set);
+  sigaddset(&set, SIGCHLD);
+  sigprocmask(SIG_BLOCK, &set, &old_set);
+}
+
 void post_hook(unsigned int argc, char* argv)
 {
-  kill(child, SIGKILL);
+  sigprocmask(SIG_SETMASK, &old_set, NULL);
+  if (child != -1) {
+    waitpid(child, NULL, WNOHANG);
+    kill(child, SIGKILL);
+  }
+}
+
+static int wait_child(void)
+{
+  sigset_t sigchild_set;
+  struct timespec timeout = {1, 0};
+
+  sigemptyset(&sigchild_set);
+  sigaddset(&sigchild_set, SIGCHLD);
+  sigtimedwait(&sigchild_set, NULL, &timeout);
+  return 0;
 }
 
 int run(int argc, char** argv)
@@ -29,7 +58,7 @@ int run(int argc, char** argv)
   if ((child = fork())) {
     if (child < 0)
       return 1;
-    return waitpid(child, NULL, 0x0) == child;
+    return wait_child();
   }
   else {
     dup2(null_fd, 0);
@@ -39,4 +68,5 @@ int run(int argc, char** argv)
     close(DEVAFL_FD);
     exit(execl(argv[1], argv[1], NULL));
   }
+
 }
